requirements.txt
bigquery-python
boto